refactor proposal operation correlation router
